feat: remove namespaces
