--- p2p_wallet/Scenes/CreateOrRestoreWallet/CreateWallet/NewCreateWallet.ViewModel.swift
@@ -1,17 +1,0 @@
-// Copyright 2022 P2P Validator Authors. All rights reserved.
-// Use of this source code is governed by a MIT-style license that can be
-// found in the LICENSE file.
-
-import Combine
-import Foundation
-import Onboarding
-
-extension NewCreateWallet {
-    final class ViewModel {
-        let onboardingStateMachine: CreateWalletStateMachine
-
-        init(tKeyFacade: TKeyFacade? = nil) {
-            onboardingStateMachine = .init(provider: tKeyFacade ?? TKeyMockupFacade())
-        }
-    }
-}
