--- p2p_wallet/Scenes/CreateOrRestoreWallet/CreateWallet/NewCreateWallet.swift
@@ -1,7 +1,0 @@
-// Copyright 2022 P2P Validator Authors. All rights reserved.
-// Use of this source code is governed by a MIT-style license that can be
-// found in the LICENSE file.
-
-import Foundation
-
-enum NewCreateWallet {}
